requirements.txt
appnope==0.1.0
argh==0.26.2
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
bokeh==1.3.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cloudpickle==1.2.1
cycler==0.10.0
dask==2.2.0
dask-glm==0.2.0
dask-ml==1.0.0
deap==1.3.0
decorator==4.4.0
defusedxml==0.6.0
distributed==2.2.0
Django==2.2.4
django-dramatiq==0.7.1
djangorestframework==3.10.2
dramatiq==1.6.1
entrypoints==0.3
fsspec==0.4.0
gevent==1.4.0
greenlet==0.4.15
HeapDict==1.0.0
idna==2.8
ipykernel==5.1.1
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.14.1
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
kiwisolver==1.1.0
llvmlite==0.29.0
locket==0.2.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
msgpack==0.6.1
multipledispatch==0.6.0
nbconvert==5.5.0
nbformat==4.4.0
notebook==6.0.0
numba==0.45.1
numpy==1.17.0
packaging==19.1
pandas==0.25.0
pandocfilters==1.4.2
parso==0.5.1
partd==1.0.0
pathtools==0.1.2
pexpect==4.7.0
pickleshare==0.7.5
pika==1.1.0
Pillow==6.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
psutil==5.6.3
psycopg2==2.8.3
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
pyzmq==18.0.2
qtconsole==4.5.2
redis==3.3.5
requests==2.22.0
scikit-learn==0.21.3
scikit-MDR==0.4.4
scipy==1.3.0
Send2Trash==1.5.0
six==1.12.0
skrebate==0.6
sortedcontainers==2.1.0
sqlparse==0.3.0
stopit==1.1.2
tblib==1.4.0
terminado==0.8.2
testpath==0.4.2
toolz==0.10.0
tornado==6.0.3
TPOT==0.10.2
tqdm==4.32.2
traitlets==4.3.2
update-checker==0.16
urllib3==1.25.3
watchdog==0.8.3
watchdog-gevent==0.1.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
xgboost==0.90
zict==1.0.0
